DTS: dialog normalization gain removal for extension substreams
Implements #2377.

+++ src/output/p_dts.cpp
@@ -100,7 +100,7 @@ dts_packetizer_c::get_dts_packet(bool flushing) {
       && dtsheader.has_core
       && dtsheader.has_exss
       && (dtsheader.exss_part_size > 0)
-      && (dtsheader.exss_part_size < static_cast<int>(dtsheader.frame_byte_size))) {
+      && (dtsheader.exss_part_size < dtsheader.frame_byte_size)) {
     dtsheader.frame_byte_size -= dtsheader.exss_part_size;
     dtsheader.has_exss         = false;
   }
